requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==23.2.0
audioread==3.0.1
boto3==1.34.128
botocore==1.34.128
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
confluent-kafka==2.3.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
fastavro==1.9.4
filelock==3.13.4
flake8==7.1.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
huggingface-hub==0.22.2
HyperPyYAML==1.2.2
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning==2.2.2
lightning-utilities==0.11.2
llvmlite==0.42.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20231117
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
protobuf==5.26.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycodestyle==2.12.0
pycparser==2.22
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.2.2
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.29
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tiktoken==0.6.0
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
yarl==1.9.4
